pyproject.toml
[project]
authors = [{name = "inwonakng", email = "inwon.kang04@gmail.com"}]
dependencies = ["numpy>=2.2.6,<3", "pandas>=2.3.2,<3", "scikit-learn>=1.7.1,<2", "openml>=0.15.1,<0.16", "ucimlrepo>=0.0.7,<0.0.8", "liac-arff>=2.5.0,<3"]
name = "tabkit"
requires-python = ">= 3.10"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
tabkit = { path = ".", editable = true }

[tool.pixi.tasks]

[tool.pixi.dependencies]
python = "3.10.*"
